Refactor knowledge browser state and dialog
